Fix g++ warning msg
git-svn-id: svn://irodssvn.ucsd.edu/trunk@3935 ea137bd6-aa86-47b6-b90d-b101d9f19a90

+++ iRODS/lib/core/src/sockComm.c
@@ -457,7 +457,7 @@ readVersion (int sock, version_t **myVersion)
           myHeader.errorLen);
     }
 
-    if (myHeader.msgLen > sizeof (version_t) * 2 || myHeader.msgLen <= 0) {
+    if (myHeader.msgLen > (int) sizeof (version_t) * 2 || myHeader.msgLen <= 0) {
         if (inputStructBBuf.buf != NULL)
             free (inputStructBBuf.buf);
         rodsLog (LOG_NOTICE,
